backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.9.0
httpx==0.27.2
gotrue==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.12.5
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0